formatting responses, adding softdeletes to movements and connecting to measurement_unit

+++ app/Models/StoreMaterial.php
@@ -31,9 +31,9 @@ class StoreMaterial extends Model
      * @var array<string, string>
      */
     protected $casts = [
-        'quantity' => 'integer',
-        'minimum_limit' => 'integer',
-        'critical_limit' => 'integer',
+        'quantity' => 'float',
+        'minimum_limit' => 'float',
+        'critical_limit' => 'float',
     ];
 
     /**
